commit 5: final push with complete working project

+++ src/main/java/com/EmployeeProject/EmployeeManagement/repository/OwnerRepository.java
@@ -3,6 +3,6 @@ package com.EmployeeProject.EmployeeManagement.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.EmployeeProject.EmployeeManagement.entity.OwnerEntity;
 
-public interface OwnerRepository extends CrudRepository<OwnerEntity, Integer> {
+public interface OwnerRepository extends CrudRepository<OwnerEntity, Long> {
 
 }
